Reorganize components and containers
