+++ src/client/containers/collection/index.jsx
