+++ src/client/containers/collection/userBadge/index.jsx
